requirement.txt
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
idna==2.6
lxml==4.1.1
Mastodon.py==1.2.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wikiquote==0.1.6
